+++ LostPeterVulkan/LostPeterEngine/src/MaterialData.cpp
@@ -132,8 +132,8 @@ namespace LostPeterEngine
     uint32 MaterialData::calculateSize() const
     {
 		uint32 nMemSize = sizeof(*this);
-		nMemSize += m_strPath.size() * sizeof(char);
-        nMemSize += GetName().size() * sizeof(char);
+		nMemSize += (uint32)(m_strPath.size() * sizeof(char));
+        nMemSize += (uint32)(GetName().size() * sizeof(char));
